refactors error and success popups into one
