+++ cypress/integration/userCanCreateArticle.feature.js
@@ -21,7 +21,7 @@ describe('User can create article', () => {
         cy.get('.item').contains('Flat Earth').click();
         cy.get('[data-cy=submit-btn]').click();
       });
-      cy.get('[data-cy=success-message]').should(
+      cy.get('[data-cy=popup-message]').should(
         'contain',
         'Your article has successfully been created'
       );
@@ -47,7 +47,7 @@ describe('User can create article', () => {
         cy.get('.item').contains('Flat Earth').click();
         cy.get('[data-cy=submit-btn]').click();
       });
-      cy.get('[data-cy=error-message]').should(
+      cy.get('[data-cy=popup-message]').should(
         'contain',
         'Something went wrong on our server, try again later'
       );
