+++ src/state/reducers/rootReducer.js
@@ -3,22 +3,22 @@ const rootReducer = (state, action) => {
     case 'SUCCESS_MESSAGE':
       return {
         ...state,
+				error: false,
         message: action.payload,
-        successOpen: true,
-      };
-    case 'CLOSE_MESSAGE':
-      return {
-        ...state,
-        successOpen: false,
-        errorOpen: false,
+        popupOpen: true,
       };
     case 'ERROR_MESSAGE':
       return {
         ...state,
-        message: action.payload,
-        errorOpen: true,
+				error: true,
+        message: action.payload,        
+        popupOpen: true,
       };
-
+    case 'CLOSE_MESSAGE':
+      return {
+        ...state,
+        popupOpen: false,
+      };
     default:
       return state;
   }
